Correcciones scrum
